third_party/WebKit/LayoutTests/fast/dom/MutationObserver/cross-document-expected.txt

Test that MutationObservers move correctly across documents

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing basic aspects of cross-document observation.
PASS mutations.length is 1
PASS mutations[0].type is "attributes"
PASS mutations[0].target is div
PASS mutations[0].attributeName is "id"
PASS mutations[0].attributeNamespace is null


Testing that subtree observation works after node is moved.
PASS mutations.length is 1
PASS mutations[0].type is "attributes"
PASS mutations[0].target is subDiv
PASS mutations[0].attributeName is "id"
PASS mutations[0].attributeNamespace is null


PASS successfullyParsed is true

TEST COMPLETE
